Rename generator to StubGenerator and package as CLI tool
