+++ src/StubGenerator/StubGenerator.csproj
@@ -5,6 +5,8 @@
     <TargetFramework>net9.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
+    <PackAsTool>true</PackAsTool>
+    <ToolCommandName>stubgen</ToolCommandName>
   </PropertyGroup>
 
   <ItemGroup>
