+++ tests/StubGenerator.Tests/StubGenerator.Tests.csproj
@@ -19,7 +19,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\src\RevitStubGenerator\RevitStubGenerator.csproj" />
+    <ProjectReference Include="..\..\src\StubGenerator\StubGenerator.csproj" />
     <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="4.9.0" />
   </ItemGroup>
